refactor: reorder code repository
